# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ISC Mbujimayi Digital Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List courses
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: filiereId
          in: query
          schema:
            type: string
        - name: teacherId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, PUBLISHED, ARCHIVED]
        - name: level
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseListResponse"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a course (Teacher only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "201":
          description: Created course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourseById
      tags: [courses]
      summary: Get course by ID with modules
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Course with modules and chapters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
    put:
      operationId: updateCourse
      tags: [courses]
      summary: Update course (Teacher only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCourseBody"
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete course (Teacher only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # Modules
  /courses/{courseId}/modules:
    get:
      operationId: listModules
      tags: [modules]
      summary: List modules for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Module"
    post:
      operationId: createModule
      tags: [modules]
      summary: Create a module (Teacher only)
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModuleBody"
      responses:
        "201":
          description: Created module
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
  /modules/{id}:
    put:
      operationId: updateModule
      tags: [modules]
      summary: Update a module
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModuleBody"
      responses:
        "200":
          description: Updated module
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
    delete:
      operationId: deleteModule
      tags: [modules]
      summary: Delete a module
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # Chapters
  /modules/{moduleId}/chapters:
    get:
      operationId: listChapters
      tags: [chapters]
      summary: List chapters for a module
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of chapters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chapter"
    post:
      operationId: createChapter
      tags: [chapters]
      summary: Create a chapter (Teacher only)
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChapterBody"
      responses:
        "201":
          description: Created chapter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
  /chapters/{id}:
    put:
      operationId: updateChapter
      tags: [chapters]
      summary: Update a chapter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChapterBody"
      responses:
        "200":
          description: Updated chapter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
    delete:
      operationId: deleteChapter
      tags: [chapters]
      summary: Delete a chapter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # Enrollments
...
components:
  schemas:
...
    Filiere:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        description:
          type: ["string", "null"]
        duration:
          type: integer
        studentCount:
          type: integer
        courseCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, code, duration, studentCount, courseCount, createdAt]
...
    Teacher:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        code:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        specialty:
          type: ["string", "null"]
        grade:
          type: ["string", "null"]
        courseCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, userId, code, firstName, lastName, courseCount, createdAt]
...
    Course:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        thumbnail:
          type: ["string", "null"]
        teacherId:
          type: string
        teacher:
          $ref: "#/components/schemas/Teacher"
        filiereId:
          type: ["string", "null"]
        filiere:
          oneOf:
            - $ref: "#/components/schemas/Filiere"
            - type: "null"
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]
        level:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        enrollmentCount:
          type: integer
        moduleCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, description, teacherId, teacher, status, enrollmentCount, moduleCount, createdAt, updatedAt]
    CourseDetail:
      allOf:
        - $ref: "#/components/schemas/Course"
        - type: object
          properties:
            modules:
              type: array
              items:
                $ref: "#/components/schemas/ModuleWithChapters"
          required: [modules]
    CourseListResponse:
      type: object
      properties:
        courses:
          type: array
          items:
            $ref: "#/components/schemas/Course"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
      required: [courses, total, page, pageSize, totalPages]
    CreateCourseBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        thumbnail:
          type: string
        filiereId:
          type: string
        level:
          type: string
        duration:
          type: integer
      required: [title, description]
    UpdateCourseBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        thumbnail:
          type: string
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]
        level:
          type: string
        duration:
          type: integer

    # Modules
    Module:
      type: object
      properties:
        id:
          type: string
        courseId:
          type: string
        title:
          type: string
        order:
          type: integer
        chapterCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, courseId, title, order, chapterCount, createdAt]
    ModuleWithChapters:
      allOf:
        - $ref: "#/components/schemas/Module"
        - type: object
          properties:
            chapters:
              type: array
              items:
                $ref: "#/components/schemas/Chapter"
          required: [chapters]
    CreateModuleBody:
      type: object
      properties:
        title:
          type: string
        order:
          type: integer
      required: [title, order]

    # Chapters
    Chapter:
      type: object
      properties:
        id:
          type: string
        moduleId:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [VIDEO, PDF, PRESENTATION, TEXT]
        content:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, moduleId, title, type, order, createdAt]
    CreateChapterBody:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [VIDEO, PDF, PRESENTATION, TEXT]
        content:
          type: string
        duration:
          type: integer
        order:
          type: integer
      required: [title, type, order]
    UpdateChapterBody:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [VIDEO, PDF, PRESENTATION, TEXT]
        content:
          type: string
        duration:
          type: integer
        order:
          type: integer

    # Enrollments
...